decoupages sont rendu en uint8 au lieu de double

+++ tp2/idecoupage.m
@@ -1,7 +1,7 @@
 function [ image ] = idecoupage( blocs, m, n )
     %on pre-alloue la matrice pour accelerer le code.
     %evite de recopier la matrice a chaque bloc lors de la boucle.
-    image = zeros(n,m);
+    image = uint8(zeros(n,m));
     
     for i=1:size(blocs, 3)
         u1 = mod((i-1)*8, m)+1;
